Move DiamondGradingReport to the appropriate module.
